+++ src/Mabipacade.Decoders/Combat/CombatActionDecoder.cs
@@ -6,6 +6,6 @@ public sealed record CombatAction;
 
 public sealed class CombatActionDecoder : IPacketDecoder
 {
-    public ushort Op => 0x7924;
+    public uint Op => 0x7924;
     public object Decode(DecoderInput input) => new CombatAction();
 }
