+++ src/Mabipacade.Decoders/Combat/CombatActionEndDecoder.cs
@@ -6,6 +6,6 @@ public sealed record CombatActionEnd;
 
 public sealed class CombatActionEndDecoder : IPacketDecoder
 {
-    public ushort Op => 0x7925;
+    public uint Op => 0x7925;
     public object Decode(DecoderInput input) => new CombatActionEnd();
 }
