+++ src/Mabipacade.Decoders/Misc/EquipmentChangedDecoder.cs
@@ -6,6 +6,6 @@ public sealed record EquipmentChanged;
 
 public sealed class EquipmentChangedDecoder : IPacketDecoder
 {
-    public ushort Op => 0x59E6;
+    public uint Op => 0x59E6;
     public object Decode(DecoderInput input) => new EquipmentChanged();
 }
