+++ src/Mabipacade.Decoders/OpCodeNames.cs
@@ -2,10 +2,10 @@ namespace Mabipacade.Decoders;
 
 public static class OpCodeNames
 {
-    private static readonly Dictionary<ushort, string> _names = Enum
+    private static readonly Dictionary<uint, string> _names = Enum
         .GetValues<OpCodes>()
-        .ToDictionary(o => (ushort)o, o => o.ToString());
+        .ToDictionary(o => (uint)o, o => o.ToString());
 
-    public static string? TryGetName(ushort op) =>
+    public static string? TryGetName(uint op) =>
         _names.TryGetValue(op, out var name) ? name : null;
 }
