+++ src/Mabipacade.Decoders/Skills/PlayerSkillPostCastAck2Decoder.cs
@@ -6,7 +6,7 @@ public sealed record PlayerSkillPostCastAck2(ushort SkillId);
 
 public sealed class PlayerSkillPostCastAck2Decoder : IPacketDecoder
 {
-    public ushort Op => 0x6989;
+    public uint Op => 0x6989;
     public object Decode(DecoderInput input) =>
         new PlayerSkillPostCastAck2(input.Elems[0].AsUInt16());
 }
