+++ src/Mabipacade.Decoders/Skills/PlayerSkillStopDecoder.cs
@@ -6,6 +6,6 @@ public sealed record PlayerSkillStop;
 
 public sealed class PlayerSkillStopDecoder : IPacketDecoder
 {
-    public ushort Op => 0x698B;
+    public uint Op => 0x698B;
     public object Decode(DecoderInput input) => new PlayerSkillStop();
 }
