+++ src/Mabipacade.Decoders/Stats/EntityRelatedDecoder.cs
@@ -6,6 +6,6 @@ public sealed record EntityRelated;
 
 public sealed class EntityRelatedDecoder : IPacketDecoder
 {
-    public ushort Op => 0x7534;
+    public uint Op => 0x7534;
     public object Decode(DecoderInput input) => new EntityRelated();
 }
