+++ src/Mabipacade.Decoders/Stats/StatUpdatePrivateDecoder.cs
@@ -6,6 +6,6 @@ public sealed record StatUpdatePrivate;
 
 public sealed class StatUpdatePrivateDecoder : IPacketDecoder
 {
-    public ushort Op => 0x7530;
+    public uint Op => 0x7530;
     public object Decode(DecoderInput input) => new StatUpdatePrivate();
 }
