+++ tests/Mabipacade.Decoders.Tests/Entity/EntitiesAppearDecoderTests.cs
@@ -9,7 +9,7 @@ public class EntitiesAppearDecoderTests
     [Fact]
     public void Op_Matches()
     {
-        Assert.Equal((ushort)0x5334, new EntitiesAppearDecoder().Op);
+        Assert.Equal((uint)0x5334, new EntitiesAppearDecoder().Op);
     }
 
     [Fact]
